♻️ Refactor Tracker Use Case to interface
For better testing, updating the dependencies from implementation to
interface.
